# Use rendering hints for zone and portrait rendering 