@Phergus 2020-04-27
This will have a big impact on map performance.  Drop a largish map image (4k x 4k or so) on a background layer and then open up the Performance Data window (Tool -> Collect Performance Data).  Zoom in and out on map.  Note rendering times.   Do the same in a release version.