@Merudo 2020-04-28
If there is a performance / quality tradeoff, then we should provide the option to the user.