@JamzTheMan 2020-04-28
Wouldn't you use the more performant option to render as you zoom and pan, and then render a final pass after x time using the higher quality?

Of course this would be more terrible to code but better than yet another option and having to choose one or the other.