@nmeier 2020-04-28
I'll run some tests on that 4k-x-4k and, say, a hundred tokens