@nmeier 2020-04-29
4500x3136 background image token adds 10-129ms to rendering, avg 37, median 34 .

While that's not much on my machine this could be noticeable on a slower machine, certainly if there are two or three of those images. 

Looking at the result of the map I'm trying out there's actually not much noticeable difference, a minor diff in jagged edge.

![image](https://user-images.githubusercontent.com/312857/80552788-b8190000-8995-11ea-94fe-2264c51cad6e.png)