@nmeier 2020-04-29
324 tokens, medium size, aligned to grid 
```
with rendering hint   , 10-111ms, 35ms avg, 26ms mean
without rendering hint, 10- 34ms, 13ms avg, 12ms median

```

So overall not much impact for this amount of tokens imo. But then again the small tokens quality is hardly noticeable.