@nmeier 2020-04-29
PointerTool's Portrait  painting with/without cubic rendering hint for image scaling 

```
without: 24-98ms, 42ms avg, 40ms median
with   : 29-87ms, 47ms avg, 42ms median

```

So overall I'd venture to say that painting the portrait in higher quality is negligible in terms of performance and renders a better picture. 
Since the AssetVieweDialog (Handout) also renders in higher quality (VALUE_INTERPOLATION_BILINEAR) the same intent of a good looking image would be in the same spirit here.