@nmeier 2020-04-29
Suggestion

1) use bilinear for portrait overlay in PointerTool - that picture will look better (especially since it's forced scaled as per user specified 'constrainTo' size). Performance is fine. Same intent as AssetViewer/Handout, same quality.

2) don't apply scaling for zone renderer (background/tokens) unless there's an option to chose high quality rendering from. Performance can be an issue depending on users' machine and irritate fluidity of moving/zooming across map.
Benefit is minimal unless you're pixel peeping at small zoomed tokens ... are small.