@Phergus 2020-04-29
Agreed.  Keep the improved scaling for portraits.  For now, leave it out of zone renderer as the minimal improvement in quality isn't worth the performance hit.